+++ src/com/blevinstein/sr/SR.java
@@ -6,27 +6,27 @@ import java.awt.geom.AffineTransform;
 import java.awt.geom.NoninvertibleTransformException;
 
 public class SR {
-  public static final float c = 10;
-  public static final float MAX = 0.95f;
+  public static final double c = 10;
+  public static final double MAX = 0.95;
 
   // given (bx, by) = beta = v / c
   // returns new spacetime coordinates for an event after transformation
   public static Event lorentz(Event e, Velocity v) {
-    float beta_sq = v.beta_sq();
-    float gamma = v.gamma();
-    float bx = v.x() / c;
-    float by = v.y() / c;
+    double beta_sq = v.beta_sq();
+    double gamma = v.gamma();
+    double bx = v.x() / c;
+    double by = v.y() / c;
     
     // avoid division by zero
     if (beta_sq == 0f) return e;
 
-    float x = - e.t() * gamma * v.x()
+    double x = - e.t() * gamma * v.x()
               + e.x() * (1 + (gamma - 1) * (bx * bx) / beta_sq)
               + e.y() * ((gamma - 1) * (bx * by) / beta_sq);
-    float y = - e.t() * gamma * by * c
+    double y = - e.t() * gamma * by * c
               + e.x() * ((gamma - 1) * (bx * by) / beta_sq)
               + e.y() * (1 + (gamma - 1) * (by * by) / beta_sq);
-    float t = e.t() * gamma
+    double t = e.t() * gamma
             - e.x() * gamma * bx / c
             - e.y() * gamma * by / c;
 
@@ -36,10 +36,10 @@ public class SR {
 
   // returns an AffineTransformation for deforming a shape
   public static AffineTransform lorentzContraction(Velocity v) {
-    float beta_sq = v.beta_sq();
-    float gamma = v.gamma();
-    float bx = v.x() / c;
-    float by = v.y() / c;
+    double beta_sq = v.beta_sq();
+    double gamma = v.gamma();
+    double bx = v.x() / c;
+    double by = v.y() / c;
 
     if (beta_sq == 0) return new AffineTransform();
